backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
SQLAlchemy==2.0.32
psycopg[binary]==3.2.1
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
reportlab==4.2.2
pandas==2.2.2
email-validator==2.2.0
argon2-cffi==23.1.0
